requirements.txt
rich==13.4.2
python-dateutil==2.8.2